Cpu mining model objects. xmrstak refactor to GPU and CPU

+++ foreman-xmrstak/src/main/java/mn/foreman/xmrstak/json/Response.java
@@ -124,7 +124,7 @@ public class Response {
 
         /** The thread rates. */
         @JsonProperty("threads")
-        public List<List<Integer>> threads;
+        public List<List<BigDecimal>> threads;
 
         /** The totals. */
         @JsonProperty("total")
